fixing some clang tidy errors

+++ include/darmok/collection.hpp
@@ -30,7 +30,7 @@ namespace darmok
                 auto& comb = combs.emplace_back();
                 for (size_t j = 0; j < n; ++j)
                 {
-                    if (i & ((uint32_t)1 << j))
+                    if (uint64_t(i) & (uint64_t(1) << j))
                     {
                         comb.insert(elms[j]);
                     }
